feat: revert to cjs
